+++ Sources/BraintreeCore/Analytics/FPTIBatchData.swift
@@ -32,6 +32,8 @@ struct FPTIBatchData: Codable {
         let endTime: Int?
         let errorDescription: String?
         let eventName: String
+        /// True if the PayPal or Venmo request is to be vaulted
+        let isVaultRequest: Bool?
         /// The type of link the SDK will be handling, currently deeplink or universal
         let linkType: String?
         /// Used for linking events from the client to server side request
@@ -46,6 +48,7 @@ struct FPTIBatchData: Codable {
             case correlationID = "correlation_id"
             case errorDescription = "error_desc"
             case eventName = "event_name"
+            case isVaultRequest = "is_vault"
             case linkType = "link_type"
             case payPalContextID = "paypal_context_id"
             case timestamp = "t"